Communications Check
Rerun System.OnStartup() if communications are reestablished after being
down.  Also, added notification of what devices are offline.
Edit global pages on tick to check if the controller is offline, or if
the HMI is also offline for the PTV, and rerun System.OnStartup() once
communications are reestablished.
Add data box 1 to both the HMI and PTV master screens as a notification
of communication loss.