Production Histogram Scale
Adjust scale of production histograms to display productions over the
total possible based off of the set rated speed.  If a production value
in a histogram is over the total rated then the histogram maximum is the
maximum production value in that histogram, otherwise the histogram
maximum is the total rated value.

Reorganize tags in Production.Histogram folder.
Create Production.Histogram.General.CurrentShift.max,
Production.Histogram.General.PreviousShift.max,
Production.Histogram.Group.CurrentShift.max,
Production.Histogram.Group.PreviousShift.max, and
Production.Histogram.Detail.Shifts.max tags.
Edit System.OnTick() and System.OnStartup() to calculate the maximum
production among all groups of machines for the current and previous
shifts.
Edit On Select and On Tick actions of the Production.Group page to
calculate the maximum production among all machines in a group for the
current and previous shifts.
Edit On Select and On Tick actions of the Production.Detail and
Developer.ProductionData screens to calculate the maximum production
among all shifts for a given machine and properly scale the histograms.
Modify the maximum of histograms 1, 2, and 3 in both groups 3 and 4 on
the Production.General and Production.Group pages.
Modify 2D movement 1 and 2, and add 2D movement 3 and 4, in group 2 on
the Production.General and Production.Group pages.