Default PTV Screen Timeout
Fix auto cycling to a screen with a timout of zero when the previous
screen has a maximum timeout of 60 seconds.

Edit the default timeout period for every main screen to be 61 seconds
instead of 60 seconds, so that if communications are working the auto
cycle logic determins the next screen before the default timout.